feat(tpldata): provide `tplroot`
Working on making SaltStack Formulas portable, find the need for a
common point of reference.  Comparing the import of `map.jinja` with and
without context:

| `tpldata` | without context    | with context             |
|-----------|--------------------|--------------------------|
| `tplfile` | template/map.jinja | template/pkg/install.sls |
| `tpldir`  | template           | template/pkg             |
| `tpldot`  | template           | template.pkg             |
| `tplroot` | template           | template                 |

With `tplroot`, it is possible to use a single point of reference for
all paths, avoiding the various incantations currently used, such as:

* `from tpldir ~ "/map.jinja" import ...`
* `from salt.file.dirname(tpldir) ~ "/map.jinja" import ...`
* `from salt['file.dirname'](tpldir) ~ "/map.jinja" import ...`

In all of the above circumstances, this simply becomes:

* `from tplroot ~ "/map.jinja" import ...`

Additional benefits are gained when referencing by colon `:` for pillars
or by period `.` for `include` statements or for any requisites.

Related issues: #10838, #41195.